nimble/host: Improve find conn by ble_addr_t
This function compares ble_addr_t and the problems starts when type 0x02
or 0x03 is used in conn

Usually it is user who should take care about the proper type, but
problem might happen when this function is called in the event of
storage overflow and we might end up with type > 0x01 in conn but we are
looking for ble_addr_t found in the storage where address is already
translated to Public or Random.

This patch fixes that.